refactor: parseColorToNumber を novelLayout に集約し underline から re-export
色パーサ (#rgb 短縮/hex 検証/任意 fallback) は underline.ts にしかなかったが、
背景色・タイトル文字色でも使うため、色/幾何の純関数置き場である novelLayout.ts へ
移動する。underline.ts は後方互換のため re-export し、既存の import・テストを維持する。
3 つ目のコピーを作らない方針 (doctrine: 重複させない)。(#273)
